Struct mezzaluna_load_path::Rubylib

pub struct Rubylib { /* private fields */ }

A Ruby load path builder that reads from the RUBYLIB environment variable.

MRI Ruby allows manipulating the require search path by setting the RUBYLIB environment variable before launching the Ruby CLI. The RUBYLIB variable is read on start-up and is expected to contain a platform-native path separator-delimited list of file system paths.

The RUBYLIB environment variable or other sequence of paths is parsed when this loader is created and is immutable. This builder is intended to be called during interpreter boot.

Paths earlier in the sequence returned from load_path have higher priority.

use std::ffi::OsStr;
use std::path::Path;
use mezzaluna_load_path::Rubylib;

// Grab the load paths from the `RUBYLIB` environment variable. If the
// variable is empty or unset, `None` is returned.
let env_loader = Rubylib::new()?;

// Search `/home/artichoke/src` first, only attempting to search
// `/usr/share/artichoke` if no file is found in `/home/artichoke/src`.
let fixed_loader = Rubylib::with_rubylib(
    OsStr::new("/home/artichoke/src:/usr/share/artichoke:./_lib"),
)?;

Implementations

impl Rubylib

pub fn new() -> Option<Self>

Create a new load path builder that reads from the RUBYLIB environment variable.

The RUBYLIB environment variable is read only once at the time this method is called. The resolved load path is immutable.

This method returns None if there are errors resolving the RUBYLIB environment variable, if the RUBYLIB environment variable is not set, or if the given RUBYLIB environment variable only contains empty paths.

Examples

use mezzaluna_load_path::Rubylib;

let loader = Rubylib::new()?;

pub fn with_rubylib(rubylib: &OsStr) -> Option<Self>

Create a new load path builder that reads from the given OsStr.

The rubylib platform string given to this method is expected to be a path string of file system paths that are delimited by the platform path separator.

The resolved load path is immutable.

This method returns None if the given rubylib argument only contains empty paths. non-empty paths.

Examples

use std::ffi::OsStr;
use mezzaluna_load_path::Rubylib;

let loader = Rubylib::with_rubylib(OsStr::new("/home/artichoke/src:/usr/share/artichoke:_lib"))?;

An empty path string returns None.

use std::ffi::OsStr;
use mezzaluna_load_path::Rubylib;

let loader = Rubylib::with_rubylib(OsStr::new(""));
assert!(loader.is_none());

let loader = Rubylib::with_rubylib(OsStr::new("::::"));
assert!(loader.is_none());

pub fn load_path(&self) -> &[PathBuf]

Return a reference to the paths in $LOAD_PATH parsed by this builder.

Because the paths in RUBYLIB have the highest priority when loading features, the returned paths should appear first in $LOAD_PATH.

Examples

use std::ffi::OsStr;
use std::path::Path;
use mezzaluna_load_path::Rubylib;

let loader = Rubylib::with_rubylib(
    OsStr::new("/home/artichoke/src:/usr/share/artichoke:_lib"),
)?;
assert_eq!(
    loader.load_path(),
    &[
        Path::new("/home/artichoke/src"),
        Path::new("/usr/share/artichoke"),
        Path::new("_lib"),
    ]
);

pub fn into_load_path(self) -> Vec<PathBuf>

Consume this loader and return its $LOAD_PATH.

Because the paths in RUBYLIB have the highest priority when loading features, the returned paths should appear first in $LOAD_PATH.

Examples

use std::ffi::OsStr;
use std::path::Path;
use mezzaluna_load_path::Rubylib;

let loader = Rubylib::with_rubylib(
    OsStr::new("/home/artichoke/src:/usr/share/artichoke:_lib"),
)?;

let load_paths = loader.into_load_path();
assert_eq!(
    load_paths,
    [
        Path::new("/home/artichoke/src"),
        Path::new("/usr/share/artichoke"),
        Path::new("_lib"),
    ]
);